requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
Django==4.2
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap-modal-forms==2.2.1
django-bootstrap4==23.1
django-userforeignkey==0.5.0
pip==23.0.1
pydantic==1.10.7
setuptools==65.5.0
soupsieve==2.4
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
